Tests and Evidence for Petcarepoliza Project
This repository contains tests for the creation of two subnets: Petcarepoliza and Petcareinfo, which are integral parts of the project. Additionally, it includes bug detection tests when deploying the precompiled contracts on the subnets. Images demonstrating the creation of a validator node in the Fuji test network, along with a subnet in that network, have been uploaded.