front-end Funcionality
The front-end code for Petcarepoliza enables users to interact with the insurance policy management system. It allows adding new policies, editing existing ones, and deleting them. It also displays a list of existing policies and provides options to connect to MetaMask and perform actions such as adding to the chain and calculating file hashes.